moved provider entity to provider_views and added support of data in 5.9
